requirements.txt
python-telegram-bot==20.7
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.6
requests==2.32.3
python-dotenv==1.0.1
stripe==9.12.0
